Add stream context runtime loading
