+++ src/Arcane.Stream.RestApi.csproj
@@ -9,7 +9,7 @@
     
     <ItemGroup>
         <PackageReference Include="SnD.Sdk" Version="1.1.0" />
-        <PackageReference Include="Arcane.Framework" Version="0.0.1-2-g1d24202" />
+        <PackageReference Include="Arcane.Framework" Version="0.0.6" />
         <PackageReference Include="System.Text.Json" Version="8.0.3" />
     </ItemGroup>
