+++ src/Models/RestApiDynamicAuthStreamContext.cs
@@ -1,18 +1,13 @@
 ﻿using System;
-using System.Text.Json.Serialization;
 using Arcane.Framework.Configuration;
 using Arcane.Framework.Sources.RestApi.Models;
 using Arcane.Stream.RestApi.Models.Base;
+using Newtonsoft.Json;
 
-namespace Arcane.Stream.RestApi.Streams.RestApiPagedFixedAuth.Configuration;
+namespace Arcane.Stream.RestApi.Models;
 
-public class RestApiPagedFixedAuthStreamContext : RestApiFixedAuthBase
+public class RestApiDynamicAuthStreamContext : RestApiDynamicAuthBase
 {
-    /// <summary>
-    /// Configuration for the page resolver.
-    /// </summary>
-    public PageResolverConfiguration PageResolverConfiguration { get; init; }
-
     /// <summary>
     /// Templated url field parameters.
     /// </summary>
@@ -52,14 +47,12 @@ public class RestApiPagedFixedAuthStreamContext : RestApiFixedAuthBase
     /// <summary>
     /// Max time to wait for rowsPerGroup to accumulate.
     /// </summary>
-    [JsonConverter(typeof(SecondsToTimeSpanConverter))]
-    [JsonPropertyName("groupingIntervalSeconds")]
     public TimeSpan GroupingInterval { get; init; }
 
     /// <summary>
     /// Data location for the sink.
     /// </summary>
-    public string SinkLocation { get; set; }
+    public string SinkLocation { get; init; }
 
     /// <summary>
     /// Number of seconds to look back when determining first set of changes to extract.
@@ -74,7 +67,6 @@ public class RestApiPagedFixedAuthStreamContext : RestApiFixedAuthBase
     /// <summary>
     /// How long to wait before polling for next result set.
     /// </summary>
-    [JsonPropertyName("changeCaptureIntervalSeconds")]
     public int ChangeCaptureInterval { get; init; }
     
     /// <summary>
