+++ src/Models/SupportedStreamKinds.cs
@@ -1,0 +1,2 @@
+﻿namespace Arcane.Stream.RestApi.Models;
+
