+++ src/Program.cs
@@ -1,12 +1,11 @@
 ﻿using System;
+using System.Reflection;
 using Arcane.Framework.Contracts;
 using Arcane.Framework.Providers;
 using Arcane.Framework.Providers.Hosting;
+using Arcane.Stream.RestApi.Models;
+using Arcane.Stream.RestApi.Models.Base;
 using Arcane.Stream.RestApi.Services;
-using Arcane.Stream.RestApi.Streams.RestApiDynamicAuth.Models;
-using Arcane.Stream.RestApi.Streams.RestApiFixedAuth.Models;
-using Arcane.Stream.RestApi.Streams.RestApiPagedDynamicAuth.Models;
-using Arcane.Stream.RestApi.Streams.RestApiPagedFixedAuth.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Serilog;
@@ -16,7 +15,6 @@ using Snd.Sdk.Storage.Base;
 using Snd.Sdk.Storage.Providers;
 using Snd.Sdk.Storage.Providers.Configurations;
 
-
 Log.Logger = DefaultLoggingProvider.CreateBootstrapLogger(nameof(Arcane));
 int exitCode;
 try
@@ -25,16 +23,18 @@ try
         .AddDatadogLogging()
         .ConfigureRequiredServices(services =>
         {
-            return services.AddStreamGraphBuilder<RestApiGraphBuilder>(context => context.StreamKind switch
+            return services.AddStreamGraphBuilder<RestApiGraphBuilder>(context =>
             {
-                "RestApiFixedAuth" => StreamContext.ProvideFromEnvironment<RestApiFixedAuthStreamContext>(),
-                "RestApiDynamicAuth" => StreamContext.ProvideFromEnvironment<RestApiDynamicAuthStreamContext>(),
-                "RestApiPagedFixedAuth" => StreamContext.ProvideFromEnvironment<RestApiPagedFixedAuthStreamContext>().LoadSecrets(),
-                "RestApiPagedDynamicAuth" => StreamContext.ProvideFromEnvironment<RestApiPagedDynamicAuthStreamContext>(),
-                _ => throw new ArgumentOutOfRangeException($"Unknown stream kind: {context.StreamKind}")
+                // Dynamically load the stream context based on the stream kind
+                // This logic is used to determine the stream context type based on the stream kind
+                // Until the StreamComponent is implemented, this .
+                var contextNS = typeof(RestApiFixedAuthStreamContext).Namespace;
+                var typeFullName = $"{contextNS}.{context.StreamKind}StreamContext";
+                var targetType = Assembly.GetExecutingAssembly().GetType(typeFullName);
+                return ((RestApiStreamContextBase)StreamContext.ProvideFromEnvironment(targetType)).LoadSecrets();
             });
         })
-     .ConfigureAdditionalServices((services, context) =>
+        .ConfigureAdditionalServices((services, context) =>
          {
              services.AddAzureBlob(AzureStorageConfiguration.CreateDefault());
              services.AddDatadogMetrics(context.ApplicationName);
